requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.6.0
asttokens 
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm
debugpy 
decorator 
distro==1.9.0
exceptiongroup
executing 
fastapi==0.115.0
Flask==3.1.0
frozenlist==1.4.1
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata 
ipykernel 
ipython 
itsdangerous==2.2.0
jedi 
Jinja2==3.1.5
jiter==0.8.2
jupyter_client 
jupyter_core 
MarkupSafe==3.0.2
matplotlib-inline 
multidict==6.1.0
nest_asyncio 
oauthlib==3.2.2
packaging 
parso 
pexpect 
pickleshare 
platformdirs 
prompt_toolkit 
proto-plus==1.25.0
protobuf==5.29.2
psutil
ptyprocess 
pure_eval 
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments 
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil 
python-dotenv==1.0.1
pyzmq 
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six 
sniffio==1.3.1
stack_data 
starlette==0.38.6
tornado
traitlets 
twilio==9.3.2
typing_extensions 
uvicorn==0.30.6
wcwidth 
websockets==13.1
Werkzeug==3.1.3
yarl==1.12.1
zipp 